requirements.txt
Django==1.6.7
42cc-pystyle
South==1.0.2
django-phonenumber-field
django-bootstrap3==6.2.2
